backend/requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
constantly==23.10.4
cryptography==42.0.5
daphne==3.0.2
Django==5.0.3
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
geographiclib==2.0
geopy==2.4.1
gevent==24.2.1
greenlet==3.0.3
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
websocket==0.2.1
websocket-client==1.7.0
zope.event==5.0
zope.interface==6.2
